requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Cors
python-dotenv
gunicorn
requests
bcrypt
pymysql
termcolor
mysql-connector-python
mysqlclient
pymysql
python-magic
Flask-Dance
validate_email
dnspython
Flask-Caching
Flask-Limiter